Reworked Prometheus board to disable/re-enable regulator
There is a design flaw on Prometheus board. When the voltage level
changes on the adjustable regulator the regulator disables. I believe it
has something to do with the fact that the feedback resistors are
telling the chip that the voltage on the output is WAY out of spec as a
result the chip will reset. In order to re-enable the chip I had to
physicall short the enable line low and then let the 10k pullup to 12V
pull the line up. Now I removed the button for the processor and tied it
into an NFET that will do this shorting/releasing processess within the
MCU.

I did notice that when I set the enable line to an output things didn't
work correctly but when I set the enable line to an input with a pull
down things worked fine. I think this problem is related with selecting
the mode of the processor on power up.